AS7-1410 Runtime OSGi Management through console
Added Framework panel that to show/edit start level
Added Sorting to Bundle list on ID, BSN and state
Moved SingleEntityToDmrBridgeImpl to framework as it is a generally
useful class for cases where a single entity is to be edited through the
framework.
Added Refresh
Added proper handling of fragment
Added 'auto activate' a setting that asks the user whether it should
automatically activate the OSGi subsystem if the user opens the runtime
view.
Modified the EntityToDmrBridgeImpl so that it include runtime attributes